example/requirements-tox.txt
openpyxl>=2.4.0
xlrd>=1.0.0
six>=1.10.0
pytest>=3.0.2
pytest-cov>=2.3.1
pytest-django>=2.9.1
mock>=2.0.0
